requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.1
django-filter==24.3
djangorestframework==3.15.2
drf-spectacular==0.27.2
idna==3.7
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
psycopg2-binary==2.9.9
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
SQLAlchemy==2.0.32
sqlparse==0.5.1
stripe==10.8.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
